chore: refactor app factory
