+++ cursus/util/__init__.py
@@ -5,3 +5,4 @@
 
 from .extensions import db  # noqa: F401
 from . import exceptions as CursusException  # noqa: F401
+from .generate_content_hash import generate_content_hash  # noqa: F401
